image-analysis-tool/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
tensorflow==2.15.0
face-recognition==1.3.0
mediapipe==0.10.9
streamlit==1.31.1
python-dotenv==1.0.0
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.8  # For YOLO object detection
